Modify simpleWatermark and simpleWatermarkTest based on comments from TimestampWatermarkTest
RB=419259
R=cbotev,mitu,ydai,ynli
A=cbotev

+++ source/src/main/java/com/linkedin/uif/source/extractor/watermark/DateWatermark.java
@@ -53,7 +53,7 @@ public class DateWatermark implements Watermark {
 
   @Override
   synchronized public HashMap<Long, Long> getIntervals(long lowWatermarkValue, long highWatermarkValue,
-      int partitionInterval, int maxIntervals) {
+      long partitionInterval, int maxIntervals) {
     HashMap<Long, Long> intervalMap = new HashMap<Long, Long>();
     final SimpleDateFormat inputFormat = new SimpleDateFormat(INPUTFORMAT);
 
@@ -68,7 +68,7 @@ public class DateWatermark implements Watermark {
     final long lowWatermark = lowWatermarkDate.getTime();
     final long highWatermark = highWatermarkDate.getTime();
 
-    int interval = this.getInterval(highWatermark - lowWatermark, partitionInterval, maxIntervals) + 1;
+    long interval = this.getInterval(highWatermark - lowWatermark, partitionInterval, maxIntervals) + 1;
     LOG.info("Recalculated partition interval:" + interval + " days");
     if (interval < 0) {
       return intervalMap;
@@ -83,7 +83,7 @@ public class DateWatermark implements Watermark {
       lwm = Long.parseLong(inputFormat.format(startTime));
 
       calendar.setTime(startTime);
-      calendar.add(Calendar.DATE, interval - 1);
+      calendar.add(Calendar.DATE, (int) (interval - 1));
       nextTime = calendar.getTime();
       hwm = Long.parseLong(inputFormat.format(nextTime.getTime() <= endTime.getTime() ? nextTime : endTime));
       intervalMap.put(lwm, hwm);
@@ -102,12 +102,12 @@ public class DateWatermark implements Watermark {
    * @param Maximum number of allowed partitions
    * @return calculated interval in days
    */
-  private int getInterval(long diffInMilliSecs, int hourInterval, int maxIntervals) {
+  private long getInterval(long diffInMilliSecs, long hourInterval, int maxIntervals) {
     if (diffInMilliSecs == 0) {
       return 0;
     }
 
-    int dayInterval = hourInterval / 24;
+    long dayInterval = hourInterval / 24;
     int totalHours = DoubleMath.roundToInt((double) diffInMilliSecs / (60 * 60 * 1000), RoundingMode.CEILING);
     int totalIntervals = DoubleMath.roundToInt((double) totalHours / (dayInterval * 24), RoundingMode.CEILING);
     if (totalIntervals > maxIntervals) {
