+++ source/src/main/java/com/linkedin/uif/source/extractor/watermark/TimestampWatermark.java
@@ -53,7 +53,7 @@ public class TimestampWatermark implements Watermark {
 
   @Override
   synchronized public HashMap<Long, Long> getIntervals(long lowWatermarkValue, long highWatermarkValue,
-      int partitionInterval, int maxIntervals) {
+      long partitionInterval, int maxIntervals) {
     HashMap<Long, Long> intervalMap = new HashMap<Long, Long>();
     final SimpleDateFormat inputFormat = new SimpleDateFormat(INPUTFORMAT);
 
@@ -66,7 +66,7 @@ public class TimestampWatermark implements Watermark {
     final long lowWatermark = this.toEpoch(Long.toString(lowWatermarkValue));
     final long highWatermark = this.toEpoch(Long.toString(highWatermarkValue));
 
-    int interval = this.getInterval(highWatermark - lowWatermark, partitionInterval, maxIntervals);
+    long interval = this.getInterval(highWatermark - lowWatermark, partitionInterval, maxIntervals);
     LOG.info("Recalculated partition interval:" + interval + " hours");
     if (interval == 0) {
       return intervalMap;
@@ -80,7 +80,7 @@ public class TimestampWatermark implements Watermark {
     while (startTime.getTime() <= endTime.getTime()) {
       lwm = Long.parseLong(inputFormat.format(startTime));
       calendar.setTime(startTime);
-      calendar.add(Calendar.HOUR, interval);
+      calendar.add(Calendar.HOUR, (int) interval);
       nextTime = calendar.getTime();
       hwm = Long.parseLong(inputFormat.format(nextTime.getTime() <= endTime.getTime() ? nextTime : endTime));
       intervalMap.put(lwm, hwm);
@@ -99,14 +99,14 @@ public class TimestampWatermark implements Watermark {
    * @param Maximum number of allowed partitions
    * @return calculated interval in hours
    */
-  private int getInterval(long diffInMilliSecs, int hourInterval, int maxIntervals) {
+  private long getInterval(long diffInMilliSecs, long hourInterval, int maxIntervals) {
     if (diffInMilliSecs == 0) {
       return 0;
     }
 
-    int totalHours = DoubleMath.roundToInt((double) diffInMilliSecs / (60 * 60 * 1000),
+    long totalHours = DoubleMath.roundToInt((double) diffInMilliSecs / (60 * 60 * 1000),
             RoundingMode.CEILING);
-    int totalIntervals = DoubleMath.roundToInt((double) totalHours / hourInterval, RoundingMode.CEILING);
+    long totalIntervals = DoubleMath.roundToInt((double) totalHours / hourInterval, RoundingMode.CEILING);
     if (totalIntervals > maxIntervals) {
       hourInterval = DoubleMath.roundToInt((double) totalHours / maxIntervals, RoundingMode.CEILING);
     }
